changed column types to date and bigint

+++ ddl-tpcds/bin_partitioned/inventory.sql
@@ -5,17 +5,17 @@ drop table if exists inventory;
 
 create table inventory
 (
-    inv_item_sk          int,
-    inv_warehouse_sk		int,
-    inv_quantity_on_hand	int
+    inv_item_sk             bigint,
+    inv_warehouse_sk        bigint,
+    inv_quantity_on_hand    bigint
 )
-partitioned by (inv_date_sk int)
+partitioned by (inv_date_sk bigint)
 stored as ${FILE};
 
 insert overwrite table inventory partition (inv_date_sk)
 select
-	i.inv_item_sk,
-	i.inv_warehouse_sk,
-	i.inv_quantity_on_hand,
-	i.inv_date_sk
+    i.inv_item_sk,
+    i.inv_warehouse_sk,
+    i.inv_quantity_on_hand,
+    i.inv_date_sk
   from ${SOURCE}.inventory i;
